controlled X values of flotsams
Lucy, 
The flotsams are now taking their xValues from a pre-defined set, stored in the flotsamXvalues array. I also changed the interval at which the flotsams are created, so it's the same as the bar, and I pushed the flotsam back, so it appears in the midddle of the bar. If you want to have more than one flotsam per bar, comment in the additional flotsams in lines 51-55.